+++ DxFramework/FrameWork/Button.cs
@@ -1,25 +1,26 @@
 ﻿using System;
 using System.Linq;
 using DxFramework.FrameWork.Bases;
+using DxFramework.FrameWork.Materials;
 using DxFramework.FrameWork.Utils;
 using DxLibDLL;
 
 namespace DxFramework.FrameWork
 {
-    internal class AutoDrawnButton : AutoDrawnText
+    internal class Button : Text
     {
-        public AutoDrawnButton(int layer)
+        public Button(int layer)
             : base(layer)
         {
             init();
         }
-        public AutoDrawnButton(int layer, Vector2 top)
+        public Button(int layer, Vector2 top)
             : base(layer)
         {
             init();
             Top = top;
         }
-        public AutoDrawnButton(int layer, Vector2 top, Action clicked)
+        public Button(int layer, Vector2 top, Action clicked)
             : base(layer)
         {
             init();
@@ -29,12 +30,12 @@ namespace DxFramework.FrameWork
                 clicked();
             };
         }
-        public AutoDrawnButton(int layer, Vector2 top, Action clicked, string text)
+        public Button(int layer, Vector2 top, Action clicked, string text)
             : base(layer)
         {
             init();
             Top = top;
-            Text.String = text;
+            TextMaterial.String = text;
             ClickedEvent += (sender, e) =>
             {
                 clicked();
@@ -45,46 +46,46 @@ namespace DxFramework.FrameWork
         {
             BackGroundFlag = true;
             Size = new Vector2(100, 50);
-            Canvas = new MultiCanvas(new Canvas(Size), new Canvas(Size), new Canvas(Size), new Canvas(Size));
+            Canvas = new MultiCanvasMaterial(new CanvasMaterial(Size), new CanvasMaterial(Size), new CanvasMaterial(Size), new CanvasMaterial(Size));
             BunedFlag = false;
             changeColor(new Color(200, 200, 200), new Color(0, 0, 0));
             setTextPosition(TextPos.Mid);
         }
 
-        public MultiCanvas Canvas
+        public MultiCanvasMaterial Canvas
         {
-            get { return base.Canvas as MultiCanvas; }
-            set { base.Canvas = value; }
+            get { return base.Material as MultiCanvasMaterial; }
+            set { base.Material = value; }
         }
 
         public bool BunedFlag { get; set; } //有効かどうか
 
-        public ICanvasBase DefaltCanvas 　// 通常時描画対象
+        public ICanvasMaterialBase DefaltCanvas 　// 通常時描画対象
         {
             get { return Canvas[0]; }
             set { Canvas[0] = value; }
         }
 
-        public ICanvasBase MousOnCanvas // マウスオーバー時描画対象
+        public ICanvasMaterialBase MousOnCanvas // マウスオーバー時描画対象
         {
             get { return Canvas[1]; }
             set { Canvas[1] = value; }
         }
 
-        public ICanvasBase ClickedCanvas // クリック時描画対象
+        public ICanvasMaterialBase ClickedCanvas // クリック時描画対象
         {
             get { return Canvas[2]; }
             set { Canvas[2] = value; }
         }
 
-        public ICanvasBase BunedCanvas // 失効時描画対象
+        public ICanvasMaterialBase BunedCanvas // 失効時描画対象
         {
             get { return Canvas[3]; }
             set { Canvas[3] = value; }
         }
 
 
-        public void setGraph(Graphic canvas)
+        public void setGraph(GraphicMaterial canvas)
         {
             int handle1 = DX.MakeGraph((int)canvas.Size.x, (int)canvas.Size.y);
             int handle2 = DX.MakeGraph((int)canvas.Size.x, (int)canvas.Size.y);
@@ -93,23 +94,23 @@ namespace DxFramework.FrameWork
             DX.GraphFilterBlt(canvas.GraphHandle, handle2, DX.DX_GRAPH_FILTER_HSB, 0, 0, 0, 100);
             DX.GraphFilterBlt(canvas.GraphHandle, handle3, DX.DX_GRAPH_FILTER_HSB, 0, 0, 0, 100);
             DefaltCanvas = canvas;
-            MousOnCanvas = new Graphic(handle1);
-            ClickedCanvas = new Graphic(handle2);
-            BunedCanvas = new Graphic(handle3);
+            MousOnCanvas = new GraphicMaterial(handle1);
+            ClickedCanvas = new GraphicMaterial(handle2);
+            BunedCanvas = new GraphicMaterial(handle3);
             Size = canvas.Size;
         }
 
         public void changeColor(Color backgroundColor,Color outlineColor)
         {
-            foreach (var itr in Canvas.OfType<Canvas>())
+            foreach (var itr in Canvas.OfType<CanvasMaterial>())
             {
                 (itr).BackGroundColor = backgroundColor;
                 (itr).OutLineColor = outlineColor;
             }
 
-            ((Canvas)MousOnCanvas).BackGroundColor += new Color(20, 20, 20);
-            ((Canvas)ClickedCanvas).BackGroundColor += new Color(50, 50, 50);
-            ((Canvas)BunedCanvas).BackGroundColor += new Color(50, 50, 50);
+            ((CanvasMaterial)MousOnCanvas).BackGroundColor += new Color(20, 20, 20);
+            ((CanvasMaterial)ClickedCanvas).BackGroundColor += new Color(50, 50, 50);
+            ((CanvasMaterial)BunedCanvas).BackGroundColor += new Color(50, 50, 50);
         }
         public void setColor(Color backgroundColor)
         {
@@ -118,12 +119,12 @@ namespace DxFramework.FrameWork
 
         public void setGraph(int handle)
         {
-            setGraph(new Graphic(handle));
+            setGraph(new GraphicMaterial(handle));
         }
 
         public void setGraph(string graphName)
         {
-            setGraph(new Graphic(graphName));
+            setGraph(new GraphicMaterial(graphName));
         }
 
         public override void update()
