+++ DxFramework/bin/Debug/Log.txt
@@ -1,99 +1,99 @@
 0:システムの情報を出力します
 1:	ＤＸライブラリ Ver3.13 
-3:	論理プロセッサの数 : 4
+1:	論理プロセッサの数 : 4
 4:	ＯＳ  Windows8 ( Build 9200  )
-111:	現時点のＣＰＵ動作速度：大体2.77GHz
+110:	現時点のＣＰＵ動作速度：大体2.79GHz
 113:	ＭＭＸ命令を使用します
 116:	ＳＳＥ命令が使用可能です
 119:	ＳＳＥ２命令が使用可能です
-121:	ＣＰＵベンダ：GenuineIntel
-133:	ＣＰＵ名：       Intel(R) Core(TM) i7-2640M CPU @ 2.80GHz
-136:ＣＯＭの初期化... 成功しました
-142:メモリ総量:16267.21MB  空きメモリ領域:10708.51MB 
-144:タイマーの精度を検査します
-147:精度結果 更新回数 マルチメディアタイマー：0  パフォーマンスカウンター：60
-150: パフォーマンスカウンターを使用します タイマー精度 : 2728.170000 KHz 
-155: ソフトの二重起動検査... 二重起動はされていませんでした
-163:ウインドウクラスを登録します... 登録に成功しました
-169:ウインドウモード起動用のウインドウを作成します
-173:ウインドウの作成に成功しました
-176:ウインドウを表示します
-184:ＩＭＥを無効にしました
-189:ウインドウスタイルをウインドウモード用に変更します... 完了
-230:DirectInput関係初期化処理
-232:	XInput DLL の読み込み中... 成功
-238:	DirectInput7 の取得中... 成功
-245:		引き続き初期化処理... 初期化成功
-251:	ジョイパッドの初期化... 
-254:	ジョイパッドの初期化は正常に終了しました
-268:	マウスデバイスの初期化... 初期化成功
-273:	キーボードデバイスの初期化... 初期化成功
-278:DirectInput 関連の初期化は正常に終了しました
-281:XAudio2 の初期化を行います
-284:	XAudio2_8.dll を使用します
-287:	XAudio2 インターフェースの取得と初期化を行います....  成功 Ver2.8
-309:XAudio2 の初期化は正常に終了しました
-313:DirectDraw オブジェクトの取得を行います....  成功
-319:引き続き初期化処理... 初期化に成功しました
-331:IDirect3D9Ex オブジェクトを取得します.... 成功
-338:IDirect3DDevice9Ex オブジェクトを取得します.... ハードウエア頂点演算を使用します
-598:成功
-600:Driver:igdumd32.dll  Description:Intel(R) HD Graphics 3000
-603:画面のフォーマットは D3DFMT_X8R8G8B8 です
-605:16bit Zバッファフォーマットは D3DFMT_D16 です
-608:24bit Zバッファフォーマットは D3DFMT_D24X8 です
-610:32bit Zバッファフォーマットは D3DFMT_D24X8 です
-613:16bit カラーフォーマットは D3DFMT_R5G6B5 です
-615:32bit カラーフォーマットは D3DFMT_X8R8G8B8 です
-617:アルファ付き 16bit カラーフォーマットは D3DFMT_A4R4G4B4 です
-619:アルファ付き 32bit カラーフォーマットは D3DFMT_A8R8G8B8 です
-621:アルファテスト用 16bit カラーフォーマットは D3DFMT_A1R5G5B5 です
-623:アルファテスト用 32bit カラーフォーマットは D3DFMT_A8R8G8B8 です
-625:ＤＸＴ１テクスチャフォーマットはD3DFMT_DXT1 です
-628:ＤＸＴ２テクスチャフォーマットはD3DFMT_DXT2 です
-630:ＤＸＴ３テクスチャフォーマットはD3DFMT_DXT3 です
-632:ＤＸＴ４テクスチャフォーマットはD3DFMT_DXT4 です
-635:ＤＸＴ５テクスチャフォーマットはD3DFMT_DXT5 です
-637:描画用 16bit カラーフォーマットは D3DFMT_R5G6B5 です
-640:描画用 32bit カラーフォーマットは D3DFMT_X8R8G8B8 です
-643:描画用アルファ付き 32bit カラーフォーマットは D3DFMT_A8R8G8B8 です
-645:描画用ABGR整数16ビット型カラーフォーマットは D3DFMT_A16B16G16R16 です
-648:描画用ABGR浮動小数点16ビット型カラーフォーマットは D3DFMT_A16B16G16R16F です
-650:描画用ABGR浮動小数点32ビット型カラーフォーマットは D3DFMT_A32B32G32R32F です
-652:描画用１チャンネル整数8ビット型カラーフォーマットは D3DFMT_A8R8G8B8 です
-655:描画用１チャンネル整数16ビット型カラーフォーマットは D3DFMT_G16R16 です
-657:描画用１チャンネル浮動小数点16ビット型カラーフォーマットは D3DFMT_R16F です
-660:描画用１チャンネル浮動小数点32ビット型カラーフォーマットは D3DFMT_R32F です
-662:描画用２チャンネル整数8ビット型カラーフォーマットは D3DFMT_A8B8G8R8 です
-664:描画用２チャンネル整数16ビット型カラーフォーマットは D3DFMT_G16R16 です
-667:描画用２チャンネル浮動小数点16ビット型カラーフォーマットは D3DFMT_G16R16F です
-669:描画用２チャンネル浮動小数点32ビット型カラーフォーマットは D3DFMT_G32R32F です
-672:一度に描画できるプリミティブの最大数:1048575
-675:対応している最大頂点インデックス:16777215
-677:同時にレンダリングできるバッファの数:4
-680:最大テクスチャサイズ　幅:8192 高さ:8192
-682:テクスチャステージテンポラリレジスタ：使用可
-685:減算合成のハードウェア対応：ネイティブ
-687:ハードウェア頂点シェーダーバージョンコード：300
-690:エミュレーション頂点シェーダーバージョンコード：エミュレーション無し
-693:ピクセルシェーダーバージョンコード：300
-716:バックバッファロック転送の時間:3330nsec  一時的な描画可能バッファを使用した転送の時間:1450nsec
-719:バックバッファロックを使用しません
-721:Ｚバッファを作成します.... 成功
-741:プログラマブルシェーダーを使用します
+122:	ＣＰＵベンダ：GenuineIntel
+161:	ＣＰＵ名：       Intel(R) Core(TM) i7-2640M CPU @ 2.80GHz
+164:ＣＯＭの初期化... 成功しました
+170:メモリ総量:16267.21MB  空きメモリ領域:10778.66MB 
+174:タイマーの精度を検査します
+176:精度結果 更新回数 マルチメディアタイマー：0  パフォーマンスカウンター：60
+179: パフォーマンスカウンターを使用します タイマー精度 : 2728.170000 KHz 
+191: ソフトの二重起動検査... 二重起動はされていませんでした
+199:ウインドウクラスを登録します... 登録に成功しました
+206:ウインドウモード起動用のウインドウを作成します
+212:ウインドウの作成に成功しました
+214:ウインドウを表示します
+222:ＩＭＥを無効にしました
+234:ウインドウスタイルをウインドウモード用に変更します... 完了
+260:DirectInput関係初期化処理
+263:	XInput DLL の読み込み中... 成功
+268:	DirectInput7 の取得中... 成功
+277:		引き続き初期化処理... 初期化成功
+283:	ジョイパッドの初期化... 
+285:	ジョイパッドの初期化は正常に終了しました
+288:	マウスデバイスの初期化... 初期化成功
+294:	キーボードデバイスの初期化... 初期化成功
+301:DirectInput 関連の初期化は正常に終了しました
+304:XAudio2 の初期化を行います
+308:	XAudio2_8.dll を使用します
+320:	XAudio2 インターフェースの取得と初期化を行います....  成功 Ver2.8
+435:XAudio2 の初期化は正常に終了しました
+440:DirectDraw オブジェクトの取得を行います....  成功
+448:引き続き初期化処理... 初期化に成功しました
+467:IDirect3D9Ex オブジェクトを取得します.... 成功
+484:IDirect3DDevice9Ex オブジェクトを取得します.... ハードウエア頂点演算を使用します
+741:成功
+754:Driver:igdumd32.dll  Description:Intel(R) HD Graphics 3000
+756:画面のフォーマットは D3DFMT_X8R8G8B8 です
+759:16bit Zバッファフォーマットは D3DFMT_D16 です
+762:24bit Zバッファフォーマットは D3DFMT_D24X8 です
+764:32bit Zバッファフォーマットは D3DFMT_D24X8 です
+767:16bit カラーフォーマットは D3DFMT_R5G6B5 です
+769:32bit カラーフォーマットは D3DFMT_X8R8G8B8 です
+773:アルファ付き 16bit カラーフォーマットは D3DFMT_A4R4G4B4 です
+776:アルファ付き 32bit カラーフォーマットは D3DFMT_A8R8G8B8 です
+778:アルファテスト用 16bit カラーフォーマットは D3DFMT_A1R5G5B5 です
+781:アルファテスト用 32bit カラーフォーマットは D3DFMT_A8R8G8B8 です
+784:ＤＸＴ１テクスチャフォーマットはD3DFMT_DXT1 です
+786:ＤＸＴ２テクスチャフォーマットはD3DFMT_DXT2 です
+788:ＤＸＴ３テクスチャフォーマットはD3DFMT_DXT3 です
+791:ＤＸＴ４テクスチャフォーマットはD3DFMT_DXT4 です
+793:ＤＸＴ５テクスチャフォーマットはD3DFMT_DXT5 です
+796:描画用 16bit カラーフォーマットは D3DFMT_R5G6B5 です
+798:描画用 32bit カラーフォーマットは D3DFMT_X8R8G8B8 です
+801:描画用アルファ付き 32bit カラーフォーマットは D3DFMT_A8R8G8B8 です
+803:描画用ABGR整数16ビット型カラーフォーマットは D3DFMT_A16B16G16R16 です
+806:描画用ABGR浮動小数点16ビット型カラーフォーマットは D3DFMT_A16B16G16R16F です
+808:描画用ABGR浮動小数点32ビット型カラーフォーマットは D3DFMT_A32B32G32R32F です
+811:描画用１チャンネル整数8ビット型カラーフォーマットは D3DFMT_A8R8G8B8 です
+813:描画用１チャンネル整数16ビット型カラーフォーマットは D3DFMT_G16R16 です
+816:描画用１チャンネル浮動小数点16ビット型カラーフォーマットは D3DFMT_R16F です
+818:描画用１チャンネル浮動小数点32ビット型カラーフォーマットは D3DFMT_R32F です
+821:描画用２チャンネル整数8ビット型カラーフォーマットは D3DFMT_A8B8G8R8 です
+823:描画用２チャンネル整数16ビット型カラーフォーマットは D3DFMT_G16R16 です
+827:描画用２チャンネル浮動小数点16ビット型カラーフォーマットは D3DFMT_G16R16F です
+829:描画用２チャンネル浮動小数点32ビット型カラーフォーマットは D3DFMT_G32R32F です
+832:一度に描画できるプリミティブの最大数:1048575
+834:対応している最大頂点インデックス:16777215
+837:同時にレンダリングできるバッファの数:4
+839:最大テクスチャサイズ　幅:8192 高さ:8192
+842:テクスチャステージテンポラリレジスタ：使用可
+844:減算合成のハードウェア対応：ネイティブ
+847:ハードウェア頂点シェーダーバージョンコード：300
+850:エミュレーション頂点シェーダーバージョンコード：エミュレーション無し
+852:ピクセルシェーダーバージョンコード：300
+875:バックバッファロック転送の時間:3452nsec  一時的な描画可能バッファを使用した転送の時間:1462nsec
+878:バックバッファロックを使用しません
+880:Ｚバッファを作成します.... 成功
+900:プログラマブルシェーダーを使用します
 
-801:フォントの初期化を行います
-804:フォントの初期化は正常に終了しました
-813:文字コードバッファの初期化を行います... 完了しました
-35411:ウインドウを閉じようとしています
-35500:ウインドウが破棄されようとしています
-35508:ソフトを終了する準備が整いました
-35519:フォントの初期化を行います
-35541:フォントの初期化は正常に終了しました
-35567:Direct3DDevice9 の解放 2
-35589:d3d9.dll の解放 1
-35592:DirectInput 関連の終了処理... 完了
-35695:
-35700:Alloc memory dump
-35706:	Total size:0(0.000kb)  Alloc num:0
-35709:
+968:フォントの初期化を行います
+972:フォントの初期化は正常に終了しました
+981:文字コードバッファの初期化を行います... 完了しました
+2972:ウインドウを閉じようとしています
+2987:ウインドウが破棄されようとしています
+3000:ソフトを終了する準備が整いました
+3023:フォントの初期化を行います
+3076:フォントの初期化は正常に終了しました
+3112:Direct3DDevice9 の解放 2
+3133:d3d9.dll の解放 1
+3139:DirectInput 関連の終了処理... 完了
+3245:
+3251:Alloc memory dump
+3255:	Total size:0(0.000kb)  Alloc num:0
+3261:
